Revert "fix: Selected To/From accounts in Filters not surfaced to top in FilterBar dropdown."
